app/requirements.txt
bleach==6.3.0
celery==5.6.2
deep-translator==1.11.4
django-admin-rangefilter==0.13.5
django-celery-results==2.6.0
django-cors-headers>=4.7.0,<5.0.0
django-filter==25.2
django-modeltranslation==0.19.19
django-translatemessages==0.0.6
django==5.2.9
djangorestframework-simplejwt==5.5.1
djangorestframework==3.16.1
drf-yasg==1.21.11
encrypted-json-field==1.1.4
factory-boy==3.3.3
flower==2.0.1
gunicorn==23.0.0
ipython==9.8.0
kid-sdk==0.8
phonenumbers==9.0.23
psycopg2-binary==2.9.11
pydantic==2.12.5
pytest-django==4.11.1
pytest-factoryboy==2.8.1
pytest-mock==3.15.1
pytest==9.0.2
requests==2.32.5
sentry-sdk==2.48.0
